Restaurants data in table with filters
Restaurants data from a fetch API request and on load displaying 5 fields and 3 filters (2 dropdowns and 1 input text with button).
Filters data based on dropdown selection or entering text and clicking enter or the button.
Once dropdown filters are set to "All"  or contents of the input fields are cleared all the data (records) are displayed again (without a need for another call)
All the filters can be used at the same instance to further filter the data.
If no matching records are found "No records to display.." message is displayed.
Basic CSS styling for a better look and no third party libraries